refactor(/app/pages/Dashboard/index): replace Cerebral with Overmind
